+++ routes/campuses/users.mjs
@@ -12,6 +12,7 @@ import {
 } from '../../models/rights';
 import User from '../../models/user';
 import config from '../../services/config';
+import contentNegociation from '../../middlewares/content-negociation';
 
 const router = new Router();
 const addDomainInError = (e) => [
@@ -22,6 +23,7 @@ const addDomainInError = (e) => [
 router.get(
   '/',
   resolveRights(CAN_LIST_CAMPUS_USER),
+  contentNegociation,
   maskOutput,
   async (ctx) => {
     const { offset, limit } = ctx.parseRangePagination(User, { max: 30 });