+++ routes/cars.mjs
@@ -3,6 +3,8 @@ import Car from '../models/car';
 import {
   CAN_CREATE_CAR, CAN_EDIT_CAR, CAN_GET_CAR, CAN_LIST_CAR, CAN_REMOVE_CAR,
 } from '../models/rights';
+import contentNegociation from '../middlewares/content-negociation';
+import maskOutput from '../middlewares/mask-output';
 
 const router = generateCRUD(Car, {
   create: {
@@ -13,6 +15,10 @@ const router = generateCRUD(Car, {
     filters: {
       campus: 'campus._id',
     },
+    middlewares: [
+      contentNegociation,
+      maskOutput,
+    ],
   },
   get: {
     right: CAN_GET_CAR,