+++ routes/pois.mjs
@@ -12,6 +12,8 @@ import {
   CAN_REMOVE_POI,
   CAN_REMOVE_POI_LOCAL,
 } from '../models/rights';
+import contentNegociation from '../middlewares/content-negociation';
+import maskOutput from '../middlewares/mask-output';
 
 const router = generateCRUD(Poi, {
   create: {
@@ -32,6 +34,10 @@ const router = generateCRUD(Poi, {
       campus: 'campus._id',
       withDisabled: 'enabled',
     },
+    middlewares: [
+      contentNegociation,
+      maskOutput,
+    ],
     async main(ctx) {
       const { offset, limit } = ctx.parseRangePagination(Poi, { max: 1000 });